Clean up and refactor TaskExecutor
Split TaskExecutor into multiple files in the 'executor' package. Also
clean up some old code.
