+++ presto-main/src/main/java/com/facebook/presto/execution/executor/TaskExecutor.java
@@ -11,20 +11,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.execution;
+package com.facebook.presto.execution.executor;
 
+import com.facebook.presto.execution.SplitRunner;
+import com.facebook.presto.execution.TaskId;
+import com.facebook.presto.execution.TaskManagerConfig;
 import com.facebook.presto.spi.PrestoException;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Ticker;
 import com.google.common.collect.ComparisonChain;
-import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.ListenableFuture;
-import com.google.common.util.concurrent.SettableFuture;
 import io.airlift.concurrent.SetThreadName;
 import io.airlift.concurrent.ThreadPoolExecutorMBean;
 import io.airlift.log.Logger;
 import io.airlift.stats.CounterStat;
-import io.airlift.stats.CpuTimer;
 import io.airlift.stats.TimeDistribution;
 import io.airlift.stats.TimeStat;
 import io.airlift.units.Duration;
@@ -37,14 +37,12 @@ import javax.annotation.concurrent.GuardedBy;
 import javax.annotation.concurrent.ThreadSafe;
 import javax.inject.Inject;
 
-import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Queue;
 import java.util.Set;
 import java.util.SortedSet;
 import java.util.concurrent.ConcurrentHashMap;
@@ -56,13 +54,11 @@ import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicLongArray;
 import java.util.function.DoubleSupplier;
 
-import static com.facebook.presto.operator.Operator.NOT_BLOCKED;
+import static com.facebook.presto.execution.executor.PrioritizedSplitRunner.calculatePriorityLevel;
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
@@ -80,17 +76,13 @@ public class TaskExecutor
 {
     private static final Logger log = Logger.get(TaskExecutor.class);
 
-    // each task is guaranteed a minimum number of tasks
+    // each task is guaranteed a minimum number of splits
     private static final int GUARANTEED_SPLITS_PER_TASK = 3;
-
-    // each time we run a split, run it for this length before returning to the pool
-    private static final Duration SPLIT_RUN_QUANTA = new Duration(1, TimeUnit.SECONDS);
 
     // print out split call stack if it has been running for a certain amount of time
     private static final Duration LONG_SPLIT_WARNING_THRESHOLD = new Duration(1000, TimeUnit.SECONDS);
 
     private static final AtomicLong NEXT_RUNNER_ID = new AtomicLong();
-    private static final AtomicLong NEXT_WORKER_ID = new AtomicLong();
 
     private final ExecutorService executor;
     private final ThreadPoolExecutorMBean executorMBean;
@@ -101,7 +93,6 @@ public class TaskExecutor
     private final Ticker ticker;
 
     private final ScheduledExecutorService splitMonitorExecutor = newSingleThreadScheduledExecutor(daemonThreadsNamed("TaskExecutor"));
-    private String currentMaxActiveSplit = null;
     private final SortedSet<RunningSplitInfo> runningSplitInfos = new ConcurrentSkipListSet<>();
 
     @GuardedBy("this")
@@ -130,7 +121,7 @@ public class TaskExecutor
     private final Set<PrioritizedSplitRunner> runningSplits = newConcurrentHashSet();
 
     /**
-     * Splits blocked by the driver (typically output buffer is full or input buffer is empty).
+     * Splits blocked by the driver.
      */
     private final Map<PrioritizedSplitRunner, Future<?>> blockedSplits = new ConcurrentHashMap<>();
 
@@ -144,10 +135,9 @@ public class TaskExecutor
     private final TimeDistribution forcedSplitWallTime = new TimeDistribution(MICROSECONDS);
 
     // shared between SplitRunners
-    private final CounterStat scheduledTimeMicros = new CounterStat();
-    private final CounterStat cpuTimeMicros = new CounterStat();
+    private final CounterStat globalCpuTimeMicros = new CounterStat();
+    private final CounterStat globalScheduledTimeMicros = new CounterStat();
 
-    private final TimeStat overallQuantaWallTime = new TimeStat(MICROSECONDS);
     private final TimeStat blockedQuantaWallTime = new TimeStat(MICROSECONDS);
     private final TimeStat unblockedQuantaWallTime = new TimeStat(MICROSECONDS);
 
@@ -173,7 +163,7 @@ public class TaskExecutor
     {
         checkArgument(runnerThreads > 0, "runnerThreads must be at least 1");
 
-        // we manages thread pool size directly, so create an unlimited pool
+        // we manage thread pool size directly, so create an unlimited pool
         this.executor = newCachedThreadPool(threadsNamed("task-processor-%s"));
         this.executorMBean = new ThreadPoolExecutorMBean((ThreadPoolExecutor) executor);
         this.runnerThreads = runnerThreads;
@@ -219,7 +209,7 @@ public class TaskExecutor
     private synchronized void addRunnerThread()
     {
         try {
-            executor.execute(new Runner());
+            executor.execute(new TaskRunner());
         }
         catch (RejectedExecutionException ignored) {
         }
@@ -272,9 +262,8 @@ public class TaskExecutor
                         taskHandle,
                         taskSplit,
                         ticker,
-                        cpuTimeMicros,
-                        scheduledTimeMicros,
-                        overallQuantaWallTime,
+                        globalCpuTimeMicros,
+                        globalScheduledTimeMicros,
                         blockedQuantaWallTime,
                         unblockedQuantaWallTime);
 
@@ -308,11 +297,11 @@ public class TaskExecutor
 
     private void splitFinished(PrioritizedSplitRunner split)
     {
-        completedSplitsPerLevel.incrementAndGet(split.priorityLevel.get());
+        completedSplitsPerLevel.incrementAndGet(split.getPriorityLevel().get());
         synchronized (this) {
             allSplits.remove(split);
 
-            long wallNanos = System.nanoTime() - split.createdNanos;
+            long wallNanos = System.nanoTime() - split.getCreatedNanos();
             splitWallTime.add(Duration.succinctNanos(wallNanos));
             if (forcedRunningSplits.remove(split)) {
                 forcedSplitWallTime.add(wallNanos);
@@ -342,7 +331,7 @@ public class TaskExecutor
             PrioritizedSplitRunner split = taskHandle.pollNextSplit();
             if (split != null) {
                 startSplit(split);
-                splitQueuedTime.add(Duration.nanosSince(split.createdNanos));
+                splitQueuedTime.add(Duration.nanosSince(split.getCreatedNanos()));
             }
         }
     }
@@ -362,7 +351,7 @@ public class TaskExecutor
                 break;
             }
 
-            splitQueuedTime.add(Duration.nanosSince(split.createdNanos));
+            splitQueuedTime.add(Duration.nanosSince(split.getCreatedNanos()));
             startSplit(split);
         }
     }
@@ -412,367 +401,14 @@ public class TaskExecutor
             }
             splitInfo.setPrinted();
 
-            currentMaxActiveSplit = splitInfo.getThreadId();
+            String currentMaxActiveSplit = splitInfo.getThreadId();
             Exception exception = new Exception("Long running split");
             exception.setStackTrace(splitInfo.getThread().getStackTrace());
             log.warn(exception, "Split thread %s has been running longer than %s", currentMaxActiveSplit, duration);
         }
     }
 
-    @ThreadSafe
-    public static class TaskHandle
-    {
-        private final TaskId taskId;
-        private final DoubleSupplier utilizationSupplier;
-        @GuardedBy("this")
-        private final Queue<PrioritizedSplitRunner> queuedSplits = new ArrayDeque<>(10);
-        @GuardedBy("this")
-        private final List<PrioritizedSplitRunner> runningSplits = new ArrayList<>(10);
-        @GuardedBy("this")
-        private final List<PrioritizedSplitRunner> forcedRunningSplits = new ArrayList<>(10);
-        @GuardedBy("this")
-        private long taskThreadUsageNanos;
-        @GuardedBy("this")
-        private boolean destroyed;
-        @GuardedBy("this")
-        private final SplitConcurrencyController concurrencyController;
-
-        private final TimeDistribution normalSplitScheduledTime;
-        private final TimeDistribution forcedSplitScheduledTime;
-
-        private final AtomicInteger nextSplitId = new AtomicInteger();
-
-        private TaskHandle(TaskId taskId, DoubleSupplier utilizationSupplier, int initialSplitConcurrency, Duration splitConcurrencyAdjustFrequency, TimeDistribution normalSplitScheduledTime, TimeDistribution forcedSplitScheduledTime)
-        {
-            this.taskId = taskId;
-            this.utilizationSupplier = utilizationSupplier;
-            this.concurrencyController = new SplitConcurrencyController(initialSplitConcurrency, splitConcurrencyAdjustFrequency);
-            this.normalSplitScheduledTime = normalSplitScheduledTime;
-            this.forcedSplitScheduledTime = forcedSplitScheduledTime;
-        }
-
-        private synchronized long addThreadUsageNanos(long durationNanos)
-        {
-            concurrencyController.update(durationNanos, utilizationSupplier.getAsDouble(), runningSplits.size());
-            taskThreadUsageNanos += durationNanos;
-            return taskThreadUsageNanos;
-        }
-
-        private TaskId getTaskId()
-        {
-            return taskId;
-        }
-
-        public synchronized boolean isDestroyed()
-        {
-            return destroyed;
-        }
-
-        // Returns any remaining splits. The caller must destroy these.
-        private synchronized List<PrioritizedSplitRunner> destroy()
-        {
-            destroyed = true;
-
-            ImmutableList.Builder<PrioritizedSplitRunner> builder = ImmutableList.builder();
-            builder.addAll(forcedRunningSplits);
-            builder.addAll(runningSplits);
-            builder.addAll(queuedSplits);
-            forcedRunningSplits.clear();
-            runningSplits.clear();
-            queuedSplits.clear();
-            return builder.build();
-        }
-
-        private synchronized void enqueueSplit(PrioritizedSplitRunner split)
-        {
-            checkState(!destroyed, "Can not add split to destroyed task handle");
-            queuedSplits.add(split);
-        }
-
-        private synchronized void recordForcedRunningSplit(PrioritizedSplitRunner split)
-        {
-            checkState(!destroyed, "Can not add split to destroyed task handle");
-            forcedRunningSplits.add(split);
-        }
-
-        @VisibleForTesting
-        synchronized int getRunningSplits()
-        {
-            return runningSplits.size();
-        }
-
-        private synchronized long getThreadUsageNanos()
-        {
-            return taskThreadUsageNanos;
-        }
-
-        private synchronized PrioritizedSplitRunner pollNextSplit()
-        {
-            if (destroyed) {
-                return null;
-            }
-
-            if (runningSplits.size() >= concurrencyController.getTargetConcurrency()) {
-                return null;
-            }
-
-            PrioritizedSplitRunner split = queuedSplits.poll();
-            if (split != null) {
-                runningSplits.add(split);
-            }
-            return split;
-        }
-
-        private synchronized void splitComplete(PrioritizedSplitRunner split)
-        {
-            concurrencyController.splitFinished(split.getSplitThreadUsageNanos(), utilizationSupplier.getAsDouble(), runningSplits.size());
-            if (forcedRunningSplits.remove(split)) {
-                forcedSplitScheduledTime.add(split.getSplitThreadUsageNanos());
-            }
-            if (runningSplits.remove(split)) {
-                normalSplitScheduledTime.add(split.getSplitThreadUsageNanos());
-            }
-        }
-
-        private int getNextSplitId()
-        {
-            return nextSplitId.getAndIncrement();
-        }
-
-        @Override
-        public String toString()
-        {
-            return toStringHelper(this)
-                    .add("taskId", taskId)
-                    .toString();
-        }
-    }
-
-    private static class PrioritizedSplitRunner
-            implements Comparable<PrioritizedSplitRunner>
-    {
-        private final long createdNanos = System.nanoTime();
-
-        private final TaskHandle taskHandle;
-        private final int splitId;
-        private final long workerId;
-        private final SplitRunner split;
-
-        private final Ticker ticker;
-
-        private final SettableFuture<?> finishedFuture = SettableFuture.create();
-
-        private final AtomicBoolean destroyed = new AtomicBoolean();
-
-        private final AtomicInteger priorityLevel = new AtomicInteger();
-        private final AtomicLong threadUsageNanos = new AtomicLong();
-        private final AtomicLong splitThreadUsageNanos = new AtomicLong();
-        private final AtomicLong lastRun = new AtomicLong();
-        private final AtomicLong start = new AtomicLong();
-
-        private final AtomicLong cpuTimeNanos = new AtomicLong();
-        private final AtomicLong processCalls = new AtomicLong();
-
-        private final CounterStat cpuTimeMicros;
-        private final CounterStat scheduledTimeMicros;
-
-        private final TimeStat overallQuantaWallTime;
-        private final TimeStat blockedQuantaWallTime;
-        private final TimeStat unblockedQuantaWallTime;
-
-        private PrioritizedSplitRunner(
-                TaskHandle taskHandle,
-                SplitRunner split,
-                Ticker ticker,
-                CounterStat cpuTimeMicros,
-                CounterStat scheduledTimeMicros,
-                TimeStat overallQuantaWallTime,
-                TimeStat blockedQuantaWallTime,
-                TimeStat unblockedQuantaWallTime)
-        {
-            this.taskHandle = taskHandle;
-            this.splitId = taskHandle.getNextSplitId();
-            this.split = split;
-            this.ticker = ticker;
-            this.workerId = NEXT_WORKER_ID.getAndIncrement();
-            this.cpuTimeMicros = cpuTimeMicros;
-            this.scheduledTimeMicros = scheduledTimeMicros;
-            this.overallQuantaWallTime = overallQuantaWallTime;
-            this.blockedQuantaWallTime = blockedQuantaWallTime;
-            this.unblockedQuantaWallTime = unblockedQuantaWallTime;
-        }
-
-        private TaskHandle getTaskHandle()
-        {
-            return taskHandle;
-        }
-
-        private ListenableFuture<?> getFinishedFuture()
-        {
-            return finishedFuture;
-        }
-
-        public boolean isDestroyed()
-        {
-            return destroyed.get();
-        }
-
-        public void destroy()
-        {
-            destroyed.set(true);
-            try {
-                split.close();
-            }
-            catch (RuntimeException e) {
-                log.error(e, "Error closing split for task %s", taskHandle.getTaskId());
-            }
-        }
-
-        public boolean isFinished()
-        {
-            boolean finished = split.isFinished();
-            if (finished) {
-                finishedFuture.set(null);
-            }
-            return finished || destroyed.get() || taskHandle.isDestroyed();
-        }
-
-        public long getSplitThreadUsageNanos()
-        {
-            return splitThreadUsageNanos.get();
-        }
-
-        public ListenableFuture<?> process()
-                throws Exception
-        {
-            try {
-                start.compareAndSet(0, ticker.read());
-
-                processCalls.incrementAndGet();
-
-                CpuTimer timer = new CpuTimer();
-                ListenableFuture<?> blocked = split.processFor(SPLIT_RUN_QUANTA);
-                CpuTimer.CpuDuration elapsed = timer.elapsedTime();
-
-                // update priority level base on total thread usage of task
-                long quantaWallNanos = elapsed.getWall().roundTo(NANOSECONDS);
-                splitThreadUsageNanos.addAndGet(quantaWallNanos);
-
-                long taskWallNanos = taskHandle.addThreadUsageNanos(quantaWallNanos);
-                threadUsageNanos.set(taskWallNanos);
-
-                priorityLevel.set(calculatePriorityLevel(taskWallNanos));
-
-                long durationMicros = elapsed.getWall().roundTo(MICROSECONDS);
-                overallQuantaWallTime.add(durationMicros, MICROSECONDS);
-
-                if (blocked == NOT_BLOCKED) {
-                    unblockedQuantaWallTime.add(durationMicros, MICROSECONDS);
-                }
-                else {
-                    blockedQuantaWallTime.add(durationMicros, MICROSECONDS);
-                }
-
-                // record last run for prioritization within a level
-                lastRun.set(ticker.read());
-
-                long cpuNanos = elapsed.getCpu().roundTo(NANOSECONDS);
-                cpuTimeNanos.addAndGet(cpuNanos);
-
-                cpuTimeMicros.update(cpuNanos / 1000);
-                scheduledTimeMicros.update(quantaWallNanos / 1000);
-
-                return blocked;
-            }
-            catch (Throwable e) {
-                finishedFuture.setException(e);
-                throw e;
-            }
-        }
-
-        public boolean updatePriorityLevel()
-        {
-            int newPriority = calculatePriorityLevel(taskHandle.getThreadUsageNanos());
-            if (newPriority == priorityLevel.getAndSet(newPriority)) {
-                return false;
-            }
-
-            // update thread usage while if level changed
-            threadUsageNanos.set(taskHandle.getThreadUsageNanos());
-            return true;
-        }
-
-        @Override
-        public int compareTo(PrioritizedSplitRunner o)
-        {
-            int level = priorityLevel.get();
-
-            int result = Integer.compare(level, o.priorityLevel.get());
-            if (result != 0) {
-                return result;
-            }
-
-            if (level < 4) {
-                result = Long.compare(threadUsageNanos.get(), o.threadUsageNanos.get());
-            }
-            else {
-                result = Long.compare(lastRun.get(), o.lastRun.get());
-            }
-            if (result != 0) {
-                return result;
-            }
-
-            return Long.compare(workerId, o.workerId);
-        }
-
-        public int getSplitId()
-        {
-            return splitId;
-        }
-
-        public String getInfo()
-        {
-            return String.format("Split %-15s-%d %s (start = %s, wall = %s ms, cpu = %s ms, calls = %s)",
-                    taskHandle.getTaskId(),
-                    splitId,
-                    split.getInfo(),
-                    start.get() / 1.0e6,
-                    (int) ((ticker.read() - start.get()) / 1.0e6),
-                    (int) (cpuTimeNanos.get() / 1.0e6),
-                    processCalls.get());
-        }
-
-        @Override
-        public String toString()
-        {
-            return String.format("Split %-15s-%d", taskHandle.getTaskId(), splitId);
-        }
-    }
-
-    private static int calculatePriorityLevel(long threadUsageNanos)
-    {
-        long millis = NANOSECONDS.toMillis(threadUsageNanos);
-
-        int priorityLevel;
-        if (millis < 1000) {
-            priorityLevel = 0;
-        }
-        else if (millis < 10_000) {
-            priorityLevel = 1;
-        }
-        else if (millis < 60_000) {
-            priorityLevel = 2;
-        }
-        else if (millis < 300_000) {
-            priorityLevel = 3;
-        }
-        else {
-            priorityLevel = 4;
-        }
-        return priorityLevel;
-    }
-
-    private class Runner
+    private class TaskRunner
             implements Runnable
     {
         private final long runnerId = NEXT_RUNNER_ID.getAndIncrement();
@@ -822,29 +458,24 @@ public class TaskExecutor
                             }
                             else {
                                 blockedSplits.put(split, blocked);
-                                blocked.addListener(new Runnable()
-                                {
-                                    @Override
-                                    public void run()
-                                    {
-                                        blockedSplits.remove(split);
-                                        split.updatePriorityLevel();
-                                        pendingSplits.put(split);
-                                    }
+                                blocked.addListener(() -> {
+                                    blockedSplits.remove(split);
+                                    split.updatePriorityLevel();
+                                    pendingSplits.put(split);
                                 }, executor);
                             }
                         }
                     }
                     catch (Throwable t) {
-                        if (split.isDestroyed()) {
-                            // ignore random errors due to driver thread interruption
-                        }
-                        else if (t instanceof PrestoException) {
-                            PrestoException e = (PrestoException) t;
-                            log.error("Error processing %s: %s: %s", split.getInfo(), e.getErrorCode().getName(), e.getMessage());
-                        }
-                        else {
-                            log.error(t, "Error processing %s", split.getInfo());
+                        // ignore random errors due to driver thread interruption
+                        if (!split.isDestroyed()) {
+                            if (t instanceof PrestoException) {
+                                PrestoException e = (PrestoException) t;
+                                log.error("Error processing %s: %s: %s", split.getInfo(), e.getErrorCode().getName(), e.getMessage());
+                            }
+                            else {
+                                log.error(t, "Error processing %s", split.getInfo());
+                            }
                         }
                         splitFinished(split);
                     }
@@ -856,57 +487,6 @@ public class TaskExecutor
                     addRunnerThread();
                 }
             }
-        }
-    }
-
-    private static class RunningSplitInfo
-            implements Comparable<RunningSplitInfo>
-    {
-        private final long startTime;
-        private final String threadId;
-        private final Thread thread;
-        private boolean printed;
-
-        public RunningSplitInfo(long startTime, String threadId, Thread thread)
-        {
-            this.startTime = startTime;
-            this.threadId = threadId;
-            this.thread = thread;
-            this.printed = false;
-        }
-
-        public long getStartTime()
-        {
-            return startTime;
-        }
-
-        public String getThreadId()
-        {
-            return threadId;
-        }
-
-        public Thread getThread()
-        {
-            return thread;
-        }
-
-        public boolean isPrinted()
-        {
-            return printed;
-        }
-
-        public void setPrinted()
-        {
-            printed = true;
-        }
-
-        @Override
-        public int compareTo(RunningSplitInfo o)
-        {
-            return ComparisonChain.start()
-                    .compare(startTime, o.getStartTime())
-                    .compare(threadId, o.getThreadId())
-                    .result();
         }
     }
 
@@ -1053,16 +633,6 @@ public class TaskExecutor
     }
 
     @Managed
-    public long getMaxActiveSplitTime()
-    {
-        Iterator<RunningSplitInfo> iterator = runningSplitInfos.iterator();
-        if (iterator.hasNext()) {
-            return NANOSECONDS.toMillis(ticker.read() - iterator.next().getStartTime());
-        }
-        return 0;
-    }
-
-    @Managed
     @Nested
     public TimeStat getSplitQueuedTime()
     {
@@ -1074,13 +644,6 @@ public class TaskExecutor
     public TimeStat getSplitWallTime()
     {
         return splitWallTime;
-    }
-
-    @Managed
-    @Nested
-    public TimeStat getOverallQuantaWallTime()
-    {
-        return overallQuantaWallTime;
     }
 
     @Managed
@@ -1127,16 +690,16 @@ public class TaskExecutor
 
     @Managed
     @Nested
-    public CounterStat getScheduledTimeMicros()
+    public CounterStat getGlobalScheduledTimeMicros()
     {
-        return scheduledTimeMicros;
+        return globalScheduledTimeMicros;
     }
 
     @Managed
     @Nested
-    public CounterStat getCpuTimeMicros()
+    public CounterStat getGlobalCpuTimeMicros()
     {
-        return cpuTimeMicros;
+        return globalCpuTimeMicros;
     }
 
     private synchronized int calculateRunningTasksForLevel(int level)
@@ -1148,6 +711,67 @@ public class TaskExecutor
             }
         }
         return count;
+    }
+
+    @Managed
+    public long getMaxActiveSplitTime()
+    {
+        Iterator<RunningSplitInfo> iterator = runningSplitInfos.iterator();
+        if (iterator.hasNext()) {
+            return NANOSECONDS.toMillis(ticker.read() - iterator.next().getStartTime());
+        }
+        return 0;
+    }
+
+    private static class RunningSplitInfo
+            implements Comparable<RunningSplitInfo>
+    {
+        private final long startTime;
+        private final String threadId;
+        private final Thread thread;
+        private boolean printed;
+
+        public RunningSplitInfo(long startTime, String threadId, Thread thread)
+        {
+            this.startTime = startTime;
+            this.threadId = threadId;
+            this.thread = thread;
+            this.printed = false;
+        }
+
+        public long getStartTime()
+        {
+            return startTime;
+        }
+
+        public String getThreadId()
+        {
+            return threadId;
+        }
+
+        public Thread getThread()
+        {
+            return thread;
+        }
+
+        public boolean isPrinted()
+        {
+            return printed;
+        }
+
+        public void setPrinted()
+        {
+            printed = true;
+        }
+
+        @Override
+        public int compareTo(RunningSplitInfo o)
+        {
+            return ComparisonChain.start()
+                    .compare(startTime, o.getStartTime())
+                    .compare(threadId, o.getThreadId())
+                    .result();
+        }
     }
 
     @Managed(description = "Task processor executor")
